+++ .gitignore
@@ -41,4 +41,4 @@ docs/design
 **/CLAUDE.md
 **/cursor.md
 
-api/uichiyomiserver
+api/uchiyomiserver
